requirements.txt
# Django
Django==1.11.10
django-environ==0.4.4
psycopg2-binary==2.7.4
mysqlclient==1.3.13
django-mysql==2.4.1

# DRF
djangorestframework==3.7.7
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-filter==1.1.0
Markdown==2.6.11

# Redis
redis==2.10.6
django-redis==4.9.0

# Celery
celery==4.2.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
flower==0.9.2

# Sentry.io
raven==6.6.0

# Development
ipdb==0.11
django-debug-toolbar==1.9.1

# AWS
boto3==1.7.75
